Relocate EEPROM drivers.
