+++ platforms/chibios/drivers/eeprom/eeprom_flash_emulated_defs.h
