+++ platforms/chibios/drivers/eeprom/eeprom_teensy.c
