+++ platforms/chibios/drivers/eeprom/eeprom_teensy.h
