+++ platforms/test/testlist.mk
@@ -1,1 +1,1 @@
-TEST_LIST += eeprom_stm32_tiny eeprom_stm32_large
+TEST_LIST += eeprom_flash_emulated_tiny eeprom_flash_emulated_large
